documentation button opens url in new tab
As long as the property "documentation" is defined either in the runtime.conf or in the BOTS file the button will open a new tab with the url provided in the property. If a new bot is added the url from BOTS is also stored in runtime.conf. IMPORTANT: the url in runtime.conf overrides the one in the BOTS file