Start replacing the PHP backend with a Python backend
The new subdirectory intelmq_manager/ contains a Python implementation
of the core functionality of intelmq-manager/php/controller.php. It
supports all of the ways in which controller.php could be called, but
the details of the URLs are a little different. Instead of one URL with
a scope parameter that indicated which of various functions to perform,
the new interface has one URL for each scope. This makes it easier to
implement and reason about the API because each of the URLs is handled
by a comparatively simple function instead of one big one.

The new backend is still incomplete in at least the following ways:

 - It's not configurable yet

 - It's doesn't add cache related header fields to the response

 - No timeout for the subprocesses

 - Needs better error handling for the subprocesses

Since the URLs have change a little, the JS code is updated to use the
new URLs.