Add Python API replacement for load_configs.php
The URL for this API is different so static.js is changed to use the
new one. The query parameters of the URL are the same as for
load_configs.php, though, so the rest the JavaScript code doesn't need
to change.

The new implementation should behave in the same way as the old one with
two exceptions: limited configurability (FILESIZE_THRESHOLD is hard
wired for now) and the new implementation doesn't automatically adapt to
IntelMQ via "intelmqctl --type json debug --get-paths".