Split FILES into WRITABLE_FILES and READONLY_FILES
Some of the configurations files defined by the FILES dictionary must be
writable by the API and some can only be read. In order to make it
easier to tell which ones are writable and so that we don't have to
spread such information about the files around, we split FILES into two
disjoint dictionaries, WRITABLE_FILES and READONLY_FILES. FILES is not
the union of these.