Add Python API replacement for save.php
The URL for this API is different so configs.js is changed to use the
new one. The query parameters of the URL are the same as for save.php,
though.

The new implementation should behave in the same way as the old one with
the exception of for now less configurability.