Add support for reading configuration from a JSON file
For now there's only one option: intelmq_ctl_cmd. The value is a list of
strings just like used by RunIntelMQCtl. Using a list of strings when
runnign the sub-process avoids going through the shell and thus avoids
shell-injections. For simplicity of the code the configuration is also a
list of strings. If this is inconvenient, we could alternatively use a
string and parse it using Python's shlex module.

This commit adds support for loading the configuration and for using it
in the intelmq_manager.api module.