Render the HTML pages with the Python backend
The Python backend now has an end-point for rendering the HTML pages.
The query parameter is the same as for the PHP backend, but the URL is a
little different.

The code is organized a little differently from the PHP version. The
information about which pages is in the templates, not in the code that
renders the templates. This keeps this information closer to the place
where it's needed, i.e. in the HTML.

This uses mako for templating. The reason for the choice of mako is that
I'm familiar with it and that Ubuntu packages are available.