Detect errors when running intelmqctl and report them to the client
Add code that detects when running intelmqctl failed and report this to
the client. Like the PHP code such errors are reported as internal
server errors with a body containing a JSON structure that gives more
information, including hints about how to solve the problem in common
cases.

Detecting these errors is not as straight forward as might be expected
because a non-zero exit code does not necessarily mean that there was an
error (e.g. intelmqctl status exits with 1 even on success when any bots
are currently stopped). This is one point in which the Python code
differs from the PHP code (see the comments in the code).