Set Cache-Control header field on the response of some API URLs
The URLs /getlog, /queues and /queues-and-status now set the
Cache-Control header field in the response. Like the old PHP
implementation the Cache-Control field uses a max-age value of 3.

The PHP implementation also set some other header fields (Expires and
Pragma) which the new API doesn't. AFAICT setting Cache-Control should
suffice and it's easy to implement.