Make file access more configurable
The restrictions of the file access API function (config and save in
api.py) were hard-wired in the code so far. This commit makes this more
flexible:

 - the parent directory for e.g. log-files can now be set in the
   configuration file (the "allowed_path" option in the file,
   ALLOWED_PATH in PHP)

 - The locations of the predefined config files that are identified in
   the API with keywords instead of filenames (e.g. "defaults", "bots",
   ...) are now automatically determined via intelmqctl.

To avoid having to modify global variables outside of the api module,
the file access code is now a class, FileAccess, which encapsulates the
configuration and whose instances can be modified in place if necessary.
In turn, there's one global instance of this class in api.py which the
api functions use and which is updated by load_api_config.