Add a "/api" prefix to the API routes
The "API routes" are the routes for all the entry points in the api.py
module except for pages which serves the HTML pages. The JavaScript used
in the HTML pages make requests to "./api/..." so when using the python
implementation directly we need to make sure the backend and the client
use the right URLs consistently.

During development I used the backend indirectly through Apache so that
I could easily use a mix of the old PHP version and the new Python
version and my Apache configuration had some URL rewriting that made the
"/api" prefix in the backend unnecessary.