Replace load_api_config with initialize_api
The new function does almost the same as the old one, but serves a
slightly different purpose. The new function is primarily meant to
initialize the API and only secondarily for loading the configuration.
The initialization needed even if one doesn't want to load a
configuration file because some of the filenames used by the file API
are determined at runtime by running intelmqctl.