Rename intelmq-manager to intelmq_manager/static
The only files remaining in the intelmq-manager directory are the static
files (js, css, etc.) so calling it static instead is more
appropriate. Also, moving it into the intelmq_manager package is useful
for python packaging purposes.

Besides the renaming, this commit also adjusts the code in
intelmq_manager/serve.py to expect the files in the new location.