Clean up the debian packaging files
The apache config was moved to the new contrib folder in the
root of the repository, because it might be also interesting for
users not using debian.
Any config related to sudo and www-data being in the `intelmq`
group has been removed. The positions.conf has been deleted, it
is now in the responsibility of the `intelmq-api` to manage that
file.
Some package metadata has been updated.
The copyright file has been updated.
The translation template for the debconf scripts has been updated
(using `debconf-updatepo`).
The debian/compat file has been dropped, it has been replaced by
a build dependency on debhelper-compat.