Remove the custom sdist command and run the html build every time
As it turns out, overriding the sdist command of setuptools is not
the best way to build the html files, because it is not used when
building a binary package- thats done without generating a source
before. Therefore the html build stage is now done in setup.py on
any run.