Add licenses for screenshots & js files & remove deprecated license file
The AGPL-3.0 is deprecated according to the reuse specification and it
is not used in any of the files in the repository, therefore was
removed.
The screenshots and the apache config snippet are insignificant and are
therefore put into public domain using the CC0 license.
The JS files were missing the license header which is fixed now.