MAINT+PKG: repl. apache virtualhost config with apache config snippet
The apache virtualhost config snipped was replaced by a simple config
snippet, that serves the intelmq-manager under the path
'/intelmq-manager' by default.
This config snippet is now installed by default into /etc/intelmq/ (like
we do with the intelma-api apache config snipped) and the Debian package
creates a softlink to /etc/apache2/conf-available and enables the
configuration on installation.
This commit also removes any Debian package files handling setup up of
the htaccess file and the relevant postrm/postinst code which was used
in the former intelmq-manager packages.