ENH: adapt the manager for the new API
IntelMQ changed the way it handles the configuration of bots, queues and
default values. IntelMQ-API followed this change and now only provides
endpoints for the runtime and the position.
This commit implements this change in the javascript code and also
updates various functions to reflect other minor changes (i.e. the bot
id is now called bot_id instead of id).